Refactor projects: * Change Renci.SshNet project into SDK-style project supporting .NET 3.5, .NET 4.0; .NET Standard 1.3, .NET Standard 2.0 and .NET Standard 2.1. * Remove .NET 3.5 project. * Remove .NET 3.5 test project. * Refactor away .NET 3.5 and .NET 4.0 specific source files. * Remove .NET Core project. * Add VS 2019 solution.
