--- src/Renci.SshNet.NETCore/Properties/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-﻿using System.Reflection;
-
-[assembly: AssemblyTitle("SSH.NET")]
