--- src/Renci.SshNet.Tests/Classes/Common/ProxyExceptionTest.NET40.cs
@@ -1,8 +1,0 @@
-﻿using Renci.SshNet.Tests.Common;
-
-namespace Renci.SshNet.Tests.Classes.Common
-{
-    public partial class ProxyExceptionTest : TestBase
-    {
-    }
-}
