+++ src/Renci.SshNet.Tests/Classes/ForwardedPortRemoteTest.cs
@@ -150,5 +150,46 @@ namespace Renci.SshNet.Tests.Classes
             ForwardedPortRemote target = new ForwardedPortRemote(boundPort, host, port);
             Assert.Inconclusive("TODO: Implement code to verify target");
         }
+
+#if FEATURE_TPL
+        [TestMethod]
+        [TestCategory("integration")]
+        public void Test_PortForwarding_Remote()
+        {
+            //  ******************************************************************
+            //  ************* Tests are still in not finished ********************
+            //  ******************************************************************
+
+            using (var client = new SshClient(Resources.HOST, Resources.USERNAME, Resources.PASSWORD))
+            {
+                client.Connect();
+                var port1 = new ForwardedPortRemote(8082, "www.renci.org", 80);
+                client.AddForwardedPort(port1);
+                port1.Exception += delegate (object sender, ExceptionEventArgs e)
+                {
+                    Assert.Fail(e.Exception.ToString());
+                };
+                port1.Start();
+                var boundport = port1.BoundPort;
+
+                System.Threading.Tasks.Parallel.For(0, 5,
+
+                    //new ParallelOptions
+                    //{
+                    //    MaxDegreeOfParallelism = 1,
+                    //},
+                    (counter) =>
+                    {
+                        var cmd = client.CreateCommand(string.Format("wget -O- http://localhost:{0}", boundport));
+                        var result = cmd.Execute();
+                        var end = DateTime.Now;
+                        System.Diagnostics.Debug.WriteLine(string.Format("Length: {0}", result.Length));
+                    }
+                );
+                Thread.Sleep(1000 * 100);
+                port1.Stop();
+            }
+        }
+#endif // FEATURE_TPL
     }
 }
